Passing all tests
Add the following logic to support linearization

Track history of put and append RPC calls via a server history to
prevent multiple calls from a client causing misordering of operations.
This order is kept by tracking each client with a random ID and each transaction
with a random ID. If there is a client put or append RPC call, the
server first checks if there is a history stored containing this clientId and transID.
If there is, we know that this operation has already occured.